refactor(cosmic-swingset): separate `exportStorage` for upcoming unit tests
